+++ src/main/java/dev/orne/beans/converters/OrneBeansConverters.java
@@ -42,6 +42,7 @@ import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.GregorianCalendar;
 import java.util.Locale;
 
 import org.apache.commons.beanutils.ConvertUtils;
@@ -72,7 +73,7 @@ import dev.orne.beans.Identity;
  *   <dt>Util converters</dt>
  *   <dd>
  *      <ul>
- *       <li>{@link CalendarConverter}</li>
+ *       <li>{@link GregorianCalendarConverter}</li>
  *       <li>{@link DateConverter}</li>
  *       <li>{@link LocaleConverter}</li>
  *      </ul>
@@ -309,11 +310,13 @@ public final class OrneBeansConverters {
     public static void registerUtilConversors(
             final boolean defaultToNull) {
         if (defaultToNull) {
-            ConvertUtils.register(new CalendarConverter((Calendar) null), Calendar.class);
+            ConvertUtils.register(new GregorianCalendarConverter((GregorianCalendar) null), GregorianCalendar.class);
+            ConvertUtils.register(new GregorianCalendarConverter((GregorianCalendar) null), Calendar.class);
             ConvertUtils.register(new DateConverter((Date) null), Date.class);
             ConvertUtils.register(new LocaleConverter((Locale) null), Locale.class);
         } else {
-            ConvertUtils.register(new CalendarConverter(), Calendar.class);
+            ConvertUtils.register(new GregorianCalendarConverter(), GregorianCalendar.class);
+            ConvertUtils.register(new GregorianCalendarConverter(), Calendar.class);
             ConvertUtils.register(new DateConverter(), Date.class);
             ConvertUtils.register(new LocaleConverter(), Locale.class);
         }
@@ -344,11 +347,13 @@ public final class OrneBeansConverters {
             final ConvertUtilsBean converter,
             final boolean defaultToNull) {
         if (defaultToNull) {
-            converter.register(new CalendarConverter((Calendar) null), Calendar.class);
+            converter.register(new GregorianCalendarConverter((GregorianCalendar) null), GregorianCalendar.class);
+            converter.register(new GregorianCalendarConverter((GregorianCalendar) null), Calendar.class);
             converter.register(new DateConverter((Date) null), Date.class);
             converter.register(new LocaleConverter((Locale) null), Locale.class);
         } else {
-            converter.register(new CalendarConverter(), Calendar.class);
+            converter.register(new GregorianCalendarConverter(), GregorianCalendar.class);
+            converter.register(new GregorianCalendarConverter(), Calendar.class);
             converter.register(new DateConverter(), Date.class);
             converter.register(new LocaleConverter(), Locale.class);
         }